updated EditProfile.js with reusuable components
