+++ src/components/Screens/profile/EditProfile/EditProfileupdated/Metrics.js
@@ -1,0 +1,10 @@
+import { Dimensions } from 'react-native';
+
+const { width, height } = Dimensions.get('window');
+
+// Responsive sizing functions
+const responsiveWidth = (size) => (width / 375) * size;
+const responsiveHeight = (size) => (height / 812) * size;
+const responsiveFont = (size) => (width / 375) * size;
+
+export { responsiveWidth, responsiveHeight, responsiveFont };
